requirements.txt
click==8.1.3
environs==9.5.0
websockets==10.3
quart==0.18.3
scheduler==0.8.0
python-telegram-bot==11.1.0